started work on SCVG download
